financial_analysis/requirements.txt
pymongo~=4.9
motor~=3.6
beanie~=1.27
fastapi~=0.115
fastapi-pagination~=0.12
fastapi-limiter~=0.1
uvicorn~=0.32
strawberry-graphql~=0.246
pydantic~=2.9
pydantic-settings~=2.6
redis~=5.1
httpx~=0.27
apscheduler~=3.10
confluent-kafka~=2.6
pandas~=2.2
numpy~=1.26
scikit-learn~=1.5